refactor(reload): separate detector state and reload responsibilities
